requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
pdfplumber==0.11.4
google-generativeai==0.8.0
python-dotenv==1.0.1